etl/requirements.txt
websocket
simplejson
websocket-client
mysql-connector-python-rf